requirements.txt
boto3
torch
transformers
Pillow
tabulate